Also detect windows-1250 every time windows-1252 is detected
This is in relation to https://github.com/aadsm/jschardet/issues/70. The differences between the two are minimal, so this will be a workaround for now. These two encodings use very different models for detection. The windows-1252 detector is purely based on the occurance probability of each character's class. The windows-1250 uses a Hungarian language model to detect the text is in Hungarian. This is brittle as there are other languages using windows-1250.